+++ src/halo_reader/scantype.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import re
 from enum import Enum, auto
 
 
@@ -21,3 +24,12 @@ class ScanType(Enum):
 
     def __repr__(self) -> str:
         return str(self)
+
+    @classmethod
+    def from_filename(cls, filename: str) -> ScanType | None:
+        match_ = re.match("^([a-zA-Z0-9]+).*", filename)
+        if match_:
+            scantype_str = match_.group(1).upper()
+            if scantype_str in cls.__members__:
+                return ScanType[scantype_str]
+        return None