+++ src/haloboard/__init__.py
@@ -1,0 +1,2 @@
+DEFAULT_ROOT = "vis"
+IMAGE_EXT = "png"