+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 ### Added
+- Added `halodata` package to iterate over halo doppler lidar data from Cloudnet
+- Added `haloboard` package for data visualisations
 
 ### Changed